feat: create extra components to increase reusability
