+++ library/lib/components/index.ts
@@ -1,4 +1,4 @@
-export * from "./Text"
+export * from "./svgs/nodes/CustomText"
 export * from "./ThemedElements"
 export * from "./Sidebar"
 export * from "./DividerLine"
